Improve help message wording
If're an admin and mess up the `users` list, you won't see "owner"
commands.

If you're just a regular room user (not an admin, not an owner), you'll
only see the `help` command in the `help` message.

Both of these situations may make you wonder:

- is that all there is?
- earlier I saw more commands, so what's going on?

Adding "and accessible to you" hopefully clears things up, or at least
it tries to make the help message more correct.